implemented fill_soap_vector_from_dscribe() without CPCtools
